Move server config modules to service
