+++ java/serving/src/main/java/feast/serving/service/config/ServingServiceConfigV2.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package feast.serving.config;
+package feast.serving.service.config;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
